// src/redux/List/list.consant.enum.ts
export enum ListEnum {
    GET_LISTS_FROM_START = 'GET_LISTS_FROM_START',
    GET_LISTS_FROM_SUCCESS = 'GET_LISTS_FROM_SUCCESS',
    GET_LISTS_FROM_FIELD = 'GET_LISTS_FROM_FIELD',

    /**
     * GIT  SINGLE  ITEM
     */
    GET_LIST_FROM_START = 'GET_LIST_FROM_START',
    GET_LIST_FROM_SUCCESS = 'GET_LIST_FROM_SUCCESS',
    GET_LIST_FROM_FIELD = 'GET_LIST_FROM_FIELD',

    /**
     *  DELATE  SINGLE  ITEM  BY  ID
     */

    DELETE_LIST_FROM_START = 'DELETE_LIST_FROM_START',
    DELETE_LIST_FROM_SUCCESS = 'DELETE_LIST_FROM_SUCCESS',
    DELETE_LIST_FROM_FIELD = 'DELETE_LIST_FROM_FIELD',

    /***  ADD  ITEM  FROM  ID   */

    ADD_LIST_FROM_START = 'ADD_LIST_FROM_START',
    ADD_LIST_FROM_SUCCESS = 'ADD_LIST_FROM_SUCCESS',
    ADD_LIST_FROM_FIELD = 'ADD_LIST_FROM_FIELD',

    GET_LIST_ID_FROM_START = 'GET_LIST_ID_FROM_START',
    GET_LIST_ID_FROM_SUCCESS = 'GET_LIST_ID_FROM_SUCCESS',
    GET_LIST_ID_FROM__FIELD = 'GET_LIST_ID_FROM__FIELD',
}